# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SessionInput": {
        "additionalProperties": false,
        "properties": {
          "learner_id": {
            "format": "uuid",
            "title": "Learner Id",
            "type": "string"
          },
          "profile_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Version Id"
          }
        },
        "required": [
          "learner_id"
        ],
        "title": "SessionInput",
        "type": "object"
      },
      "SessionPublic": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "learner_id": {
            "format": "uuid",
            "title": "Learner Id",
            "type": "string"
          },
          "problems": {
            "items": {
              "$ref": "#/components/schemas/ProblemPublic"
            },
            "title": "Problems",
            "type": "array"
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileSettings"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "learner_id",
          "status",
          "profile",
          "problems"
        ],
        "title": "SessionPublic",
        "type": "object"
      },
...
      "SessionSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "learner_id": {
            "format": "uuid",
            "title": "Learner Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "learner_id",
          "status",
          "created_at"
        ],
        "title": "SessionSummary",
        "type": "object"
      },
...
      "TutorRequestReceiptPublic": {
        "properties": {
          "activity_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Id"
          },
          "kind": {
            "enum": [
              "session",
              "activity",
              "submission"
            ],
            "title": "Kind",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "submission_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Id"
          }
        },
        "required": [
          "kind",
          "session_id"
        ],
        "title": "TutorRequestReceiptPublic",
        "type": "object"
      },
      "TutorRequestResolutionInput": {
        "additionalProperties": false,
        "properties": {
          "learner_id": {
            "format": "uuid",
            "title": "Learner Id",
            "type": "string"
          }
        },
        "required": [
          "learner_id"
        ],
        "title": "TutorRequestResolutionInput",
        "type": "object"
      },
      "TutorRequestResolutionPublic": {
        "properties": {
          "receipt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TutorRequestReceiptPublic"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "accepted",
              "not_accepted"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "TutorRequestResolutionPublic",
        "type": "object"
      },
...
      "TutoringSessionPublic": {
        "properties": {
          "difficulty": {
            "enum": [
              "introductory",
              "standard",
              "challenge"
            ],
            "title": "Difficulty",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "initiative": {
            "enum": [
              "tutor_led",
              "balanced",
              "learner_led"
            ],
            "title": "Initiative",
            "type": "string"
          },
          "learner_id": {
            "format": "uuid",
            "title": "Learner Id",
            "type": "string"
          },
          "problems": {
            "items": {
              "$ref": "#/components/schemas/ProblemPublic"
            },
            "title": "Problems",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "id",
          "learner_id",
          "status",
          "topic",
          "initiative",
          "difficulty",
          "problems"
        ],
        "title": "TutoringSessionPublic",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Shepherd Academy Universe API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/sessions": {
      "get": {
        "operationId": "sessions_api_v1_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionSummary"
                  },
                  "title": "Response Sessions Api V1 Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Sessions",
        "tags": [
          "practice"
        ]
      },
      "post": {
        "operationId": "start_session_api_v1_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Session",
        "tags": [
          "practice"
        ]
      }
    },
    "/api/v1/sessions/{session_id}": {
      "get": {
        "operationId": "read_session_api_v1_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Session",
        "tags": [
          "practice"
        ]
      }
    },
...
    "/api/v1/tutor/request-receipts/{key}": {
      "get": {
        "operationId": "request_receipt_api_v1_tutor_request_receipts__key__get",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "learner_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutorRequestReceiptPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Receipt",
        "tags": [
          "AI tutoring"
        ]
      }
    },
    "/api/v1/tutor/request-receipts/{key}/resolve": {
      "post": {
        "operationId": "resolve_request_api_v1_tutor_request_receipts__key__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorRequestResolutionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutorRequestResolutionPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Request",
        "tags": [
          "AI tutoring"
        ]
      }
    },
    "/api/v1/tutor/sessions": {
      "get": {
        "operationId": "sessions_api_v1_tutor_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TutoringSessionPublic"
                  },
                  "title": "Response Sessions Api V1 Tutor Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Sessions",
        "tags": [
          "AI tutoring"
        ]
      },
      "post": {
        "operationId": "create_session_api_v1_tutor_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutoringSessionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutoringSessionPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "AI tutoring"
        ]
      }
    },
    "/api/v1/tutor/sessions/{session_id}": {
      "get": {
        "operationId": "read_session_api_v1_tutor_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutoringSessionPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Session",
        "tags": [
          "AI tutoring"
        ]
      }
    },
    "/api/v1/tutor/sessions/{session_id}/activities": {
      "post": {
        "operationId": "activity_api_v1_tutor_sessions__session_id__activities_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorActivityInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activity",
        "tags": [
          "AI tutoring"
        ]
      }
    },
    "/api/v1/tutor/sessions/{session_id}/settings": {
      "post": {
        "operationId": "settings_api_v1_tutor_sessions__session_id__settings_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorSettingsInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutoringSessionPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Settings",
        "tags": [
          "AI tutoring"
        ]
      }
    },
...
  }
}